WIP basic version of WebOfScience::Client class
This pulls from Peter's script and basically pulls up the various
components that went into the Sciencewire client and request classes.

Once fully conformant to the new API, including limits and such, I
believe this class signature would be sufficient to stand in for the existing
ScienceWire client (and some related classes).