Use XML layout/template for WOS requests bodies
You can see now why it makes more sense for class that is using rails
templates in `app/views` to live in `app` and not `lib`.

Added `net_http_persistent` for keep-alive performance.

Note: JSON is not an available format anymore, so those params needed to
be removed.

Next step: work on the search XML template/building.