CI was running twice because of system fork out of Rakefile
Reduce cruft, improve accuracy.

The purpose was to avoid running CI in production, but that can be
accomplished without a fork.

Do `db:setup` in `before_script`.  Now the `Rails.env.production?` check
is unnecessary, because `RSpec` won't even be part of the bundle on a
production system, so all it would do is run rubocop then fail on a
NameError.

This incorporates the changes from @mjgiarlo's gist:
https://gist.github.com/mjgiarlo/ee079cd476b94066886f6f541e624a5b

And resolves the double-running of spec and rubocop.

NOTE: Explicitly (re)defining the `:spec` task was part of the mechanism
of duplication.