Move ScienceWireHarvester internal methods to private
Distinguishing between the public API of a class and the internals is
important to logic and refactorability.

All these methods are not invoked anywhere outside the class at runtime.

In one case, I revised the spec to call `send` where the method is now
private.

Also reshuffle private methods to alphanum order.