Publication model has methods that should be private
NOTE: for the purposes of this revision, I am treating
`cap_cutover.rake` as a vestigial/deprecated piece of code. Where it is
the sole (external) caller for a given method, I am not treating that as
a continuing "current" use.  I edited it so the task will still work the
same (by using `.send`, similar to the specs), but I am allowing `Publication` to take the
otherwise private method private.