sul_pub-368 WebOfScience::Record#to_pub_hash
-Refactored Record toward the SRP pattern by moving out all the logic of
WOS XML Document information extraction, in its own class named
Wos_Extractor.

-Record now simply act as the pure data type that it is, that is a
Record. It provide methods to access the Record relevant information and
the toPubHash Method to get a PubHash from it. Ideally that function
should be moved in the consumer of the Record. The record can't all its
consumer in advance. But a consumer as the PubHash class knows, that it
can be initialized from.

-As of now we extract 90% of what is required. In comment are the
contentious element that are missing. The other fields missing are the 2
PubMed fields that can be extracted when the Record is a MEDLINE Record.
This can be determine by the Database field.