Use associations, better hygiene around saving, including dependents
Clarified responsibilities and timing around saving, especially for
methods in the callback chain.

The `.save!` in the API are useful because the `.reload` in the next
line is going to fail anyway if the save didn't succeed!

In `Publication`, the main associations have `autosave: true`,
which means we shouldn't take separate trips to the DB
to save dependent objects, but instead tie their persistence to the
`.save` call on the `Publication` object.  See:
http://api.rubyonrails.org/classes/ActiveRecord/AutosaveAssociation.html

Consolidated some of the private sync functionality in Publication.

Use ActiveRecord associations to build/initialize/delete dependent
objects.