Rework cap_profile_id_rewriter away from complete hackery
DISABLES the class from instantiation, per agreement.  Further salvage
work deferred.

- Only one method here is ever invoked. Make everything else private.
- Remove unnecessary `dotiw` dependency. The `DateHelper` is good enough.
- Completely avoid all the kludgey instance variables and initialization.
- Cut out all the logging of variables that NEVER CHANGE. They are uninformative.
- Cut out the DUPLICATE log lines.
- Add YARD
- Ensure logs get written, even after a `rescue`
- Consolidate `Author` finding logic
- Add `!` to critical operations

We have a blocker in production data to the **existing** logic of this
class, specifically documented in #546.

Ironically, this class that complicates making `cap_profile_id` unique,
also requires `cap_profile_id` to be present AND requires 3 other fields
to be unique.