Add a validator for identifier_type presence
Obviously, the pub ID record is useless if it doesn't identify what kind
of ID it is.

We would want to add validation for `identifier_value` as well, but, ironically,
that is blocked because it breaks dozens of lines of specs for code designed to
clean up Pub IDs.  So the code meant to clean up values is what prevents
us from ensuring values are even there.  :\

TODO: We could also choose to enumerate valid values for type.