Add a distinct method for pairing author to wos_uid
The expected behavior is different than `process_uids` because of how
the matching against existing pubs/contribs is treated.  For
`process_uids`, matches remove the wos_uid from the input, effectively
treating it as a non-success.  This new method is satisfied by that
match, and returns the Publication itself.

Include tests.