The fix
The main strength here is that during harvest, the Pub, Contrib and the
linkage to the source record happen at **EXACTLY THE SAME TIME** and
inside a DB transaction such that it is harder for things to fall out of
joint (a historical problem w/ this app).

Backfilling of Pubs and Contribs is now centralized to the `Author` model.

We stop testing for `Identifiers` to fail because (1) it only raises in
one obvious early condition and (2) we don't even call it as much because we can
pass the *existing* record object in (with existing Identifiers).